channels: fix drdynvc, patch memory leaks
