+++ channels/drdynvc/client/drdynvc_main.h
@@ -21,11 +21,76 @@
 #define __DRDYNVC_MAIN_H
 
 #include <winpr/wlog.h>
+#include <winpr/synch.h>
+#include <freerdp/settings.h>
+#include <winpr/collections.h>
 
 #include <freerdp/api.h>
 #include <freerdp/svc.h>
+#include <freerdp/dvc.h>
 #include <freerdp/addin.h>
+#include <freerdp/channels/log.h>
 #include <freerdp/client/drdynvc.h>
+
+typedef struct drdynvc_plugin drdynvcPlugin;
+
+#define MAX_PLUGINS 32
+
+struct _DVCMAN
+{
+	IWTSVirtualChannelManager iface;
+
+	drdynvcPlugin* drdynvc;
+
+	int num_plugins;
+	const char* plugin_names[MAX_PLUGINS];
+	IWTSPlugin* plugins[MAX_PLUGINS];
+
+	int num_listeners;
+	IWTSListener* listeners[MAX_PLUGINS];
+
+	wArrayList* channels;
+	wStreamPool* pool;
+};
+typedef struct _DVCMAN DVCMAN;
+
+struct _DVCMAN_LISTENER
+{
+	IWTSListener iface;
+
+	DVCMAN* dvcman;
+	char* channel_name;
+	UINT32 flags;
+	IWTSListenerCallback* listener_callback;
+};
+typedef struct _DVCMAN_LISTENER DVCMAN_LISTENER;
+
+struct _DVCMAN_ENTRY_POINTS
+{
+	IDRDYNVC_ENTRY_POINTS iface;
+
+	DVCMAN* dvcman;
+	ADDIN_ARGV* args;
+	rdpSettings* settings;
+};
+typedef struct _DVCMAN_ENTRY_POINTS DVCMAN_ENTRY_POINTS;
+
+struct _DVCMAN_CHANNEL
+{
+	IWTSVirtualChannel iface;
+
+	int status;
+	DVCMAN* dvcman;
+	void* pInterface;
+	UINT32 channel_id;
+	char* channel_name;
+	IWTSVirtualChannelCallback* channel_callback;
+
+	wStream* dvc_data;
+	UINT32 dvc_data_length;
+	CRITICAL_SECTION lock;
+};
+typedef struct _DVCMAN_CHANNEL DVCMAN_CHANNEL;
 
 enum _DRDYNVC_STATE
 {
@@ -54,7 +119,7 @@ struct drdynvc_plugin
 	wStream* data_in;
 	void* InitHandle;
 	DWORD OpenHandle;
-	wMessagePipe* MsgPipe;
+	wMessageQueue* queue;
 
 	DRDYNVC_STATE state;
 	DrdynvcClientContext* context;
@@ -68,7 +133,6 @@ struct drdynvc_plugin
 
 	IWTSVirtualChannelManager* channel_mgr;
 };
-typedef struct drdynvc_plugin drdynvcPlugin;
 
 int drdynvc_write_data(drdynvcPlugin* plugin, UINT32 ChannelId, BYTE* data, UINT32 data_size);
