--- channels/drdynvc/client/drdynvc_types.h
@@ -1,38 +1,0 @@
-/**
- * FreeRDP: A Remote Desktop Protocol Implementation
- * Dynamic Virtual Channel
- *
- * Copyright 2010-2011 Vic Lee
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#ifndef __DRDYNVC_TYPES_H
-#define __DRDYNVC_TYPES_H
-
-#ifdef HAVE_CONFIG_H
-#include "config.h"
-#endif
-
-#include <freerdp/dvc.h>
-#include <freerdp/types.h>
-#include <freerdp/channels/log.h>
-
-#define TAG CHANNELS_TAG("dvcman.client")
-#ifdef WITH_DEBUG_DVC
-#define DEBUG_DVC(fmt, ...) WLog_DBG(TAG, fmt, ## __VA_ARGS__)
-#else
-#define DEBUG_DVC(fmt, ...) do { } while (0)
-#endif
-
-#endif
