+++ channels/remdesk/client/remdesk_main.h
@@ -41,12 +41,14 @@ struct remdesk_plugin
 	CHANNEL_DEF channelDef;
 	CHANNEL_ENTRY_POINTS_FREERDP channelEntryPoints;
 
+	RemdeskClientContext* context;
+
 	HANDLE thread;
 	wStream* data_in;
 	void* InitHandle;
 	DWORD OpenHandle;
-	wMessagePipe* MsgPipe;
 	rdpSettings* settings;
+	wMessageQueue* queue;
 
 	UINT32 Version;
 	char* ExpertBlob;
