+++ libfreerdp/core/client.c
@@ -85,7 +85,7 @@ rdpChannels* freerdp_channels_new(void)
 	if (!channels)
 		return NULL;
 
-	channels->MsgPipe = MessagePipe_New();
+	channels->queue = MessageQueue_New(NULL);
 
 	if (!g_OpenHandles)
 	{
@@ -98,20 +98,10 @@ rdpChannels* freerdp_channels_new(void)
 
 void freerdp_channels_free(rdpChannels* channels)
 {
-	int index;
-	CHANNEL_OPEN_DATA* pChannelOpenData;
-
-	MessagePipe_Free(channels->MsgPipe);
-
-	for (index = 0; index < channels->clientDataCount; index++)
+	if (channels->queue)
 	{
-		pChannelOpenData = &channels->openDataList[index];
-
-		if (pChannelOpenData->pInterface)
-		{
-			free(pChannelOpenData->pInterface);
-			pChannelOpenData->pInterface = NULL;
-		}
+		MessageQueue_Free(channels->queue);
+		channels->queue = NULL;
 	}
 
 	free(channels);
@@ -281,7 +271,7 @@ static int freerdp_channels_process_sync(rdpChannels* channels, freerdp* instanc
 	CHANNEL_OPEN_EVENT* item;
 	CHANNEL_OPEN_DATA* pChannelOpenData;
 
-	while (MessageQueue_Peek(channels->MsgPipe->Out, &message, TRUE))
+	while (MessageQueue_Peek(channels->queue, &message, TRUE))
 	{
 		if (message.id == WMQ_QUIT)
 		{
@@ -311,13 +301,6 @@ static int freerdp_channels_process_sync(rdpChannels* channels, freerdp* instanc
 
 			free(item);
 		}
-		else if (message.id == 1)
-		{
-			/**
-			 * Ignore for now, the same event is being pushed on the In queue,
-			 * and we're pushing it on the Out queue just to wake other threads
-			 */
-		}
 	}
 
 	return status;
@@ -331,7 +314,7 @@ BOOL freerdp_channels_get_fds(rdpChannels* channels, freerdp* instance, void** r
 {
 	void* pfd;
 
-	pfd = GetEventWaitObject(MessageQueue_Event(channels->MsgPipe->Out));
+	pfd = GetEventWaitObject(MessageQueue_Event(channels->queue));
 
 	if (pfd)
 	{
@@ -361,7 +344,7 @@ HANDLE freerdp_channels_get_event_handle(freerdp* instance)
 	rdpChannels* channels;
 
 	channels = instance->context->channels;
-	event = MessageQueue_Event(channels->MsgPipe->Out);
+	event = MessageQueue_Event(channels->queue);
 
 	return event;
 }
@@ -372,7 +355,7 @@ int freerdp_channels_process_pending_messages(freerdp* instance)
 
 	channels = instance->context->channels;
 
-	if (WaitForSingleObject(MessageQueue_Event(channels->MsgPipe->Out), 0) == WAIT_OBJECT_0)
+	if (WaitForSingleObject(MessageQueue_Event(channels->queue), 0) == WAIT_OBJECT_0)
 	{
 		 return freerdp_channels_process_sync(channels, instance);
 	}
@@ -385,7 +368,7 @@ int freerdp_channels_process_pending_messages(freerdp* instance)
  */
 BOOL freerdp_channels_check_fds(rdpChannels* channels, freerdp* instance)
 {
-	if (WaitForSingleObject(MessageQueue_Event(channels->MsgPipe->Out), 0) == WAIT_OBJECT_0)
+	if (WaitForSingleObject(MessageQueue_Event(channels->queue), 0) == WAIT_OBJECT_0)
 	{
 		freerdp_channels_process_sync(channels, instance);
 	}
@@ -409,10 +392,6 @@ void freerdp_channels_close(rdpChannels* channels, freerdp* instance)
 		ChannelDisconnectedEventArgs e;
 
 		pChannelClientData = &channels->clientDataList[index];
-
-		if (pChannelClientData->pChannelInitEventProc)
-			pChannelClientData->pChannelInitEventProc(pChannelClientData->pInitHandle, CHANNEL_EVENT_TERMINATED, 0, 0);
-
 		pChannelOpenData = &channels->openDataList[index];
 
 		name = (char*) malloc(9);
@@ -425,10 +404,12 @@ void freerdp_channels_close(rdpChannels* channels, freerdp* instance)
 		PubSub_OnChannelDisconnected(instance->context->pubSub, instance->context, &e);
 
 		free(name);
+
+		if (pChannelClientData->pChannelInitEventProc)
+			pChannelClientData->pChannelInitEventProc(pChannelClientData->pInitHandle, CHANNEL_EVENT_TERMINATED, 0, 0);
 	}
 
-	/* Emit a quit signal to the internal message pipe. */
-	MessagePipe_PostQuit(channels->MsgPipe, 0);
+	MessageQueue_PostQuit(channels->queue, 0);
 }
 
 UINT VCAPITYPE FreeRDP_VirtualChannelInit(LPVOID* ppInitHandle, PCHANNEL_DEF pChannel,
@@ -614,7 +595,7 @@ UINT VCAPITYPE FreeRDP_VirtualChannelWrite(DWORD openHandle, LPVOID pData, ULONG
 	pChannelOpenEvent->UserData = pUserData;
 	pChannelOpenEvent->pChannelOpenData = pChannelOpenData;
 
-	MessageQueue_Post(channels->MsgPipe->Out, (void*) channels, 0, (void*) pChannelOpenEvent, NULL);
+	MessageQueue_Post(channels->queue, (void*) channels, 0, (void*) pChannelOpenEvent, NULL);
 
 	return CHANNEL_RC_OK;
 }
@@ -627,7 +608,7 @@ int freerdp_channels_client_load(rdpChannels* channels, rdpSettings* settings, v
 
 	if (channels->clientDataCount + 1 >= CHANNEL_MAX_COUNT)
 	{
-		WLog_ERR(TAG,  "error: too many channels");
+		WLog_ERR(TAG, "error: too many channels");
 		return 1;
 	}
